fix: references service cannot parse month field
The recommended format for `month` in biblatex is to use macro,
i.e. month=jan. The current implementation of references service
matches the `rightValue` (what kind of genius wrote such code?)
the value of each field by regex looking for `"`(quote) and `{`
(brackets). The patch implemented a very *ugly* fix to remove all
`month.*,` from input, so as to fix this issue.